services/nlp_py/api_requirements.txt
# API Server Dependencies
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1

# Core RSS and Translation Dependencies
feedparser==6.0.10
deep-translator==1.11.4  # Free Google Translate API wrapper
google-cloud-translate==3.21.1  # Optional: official Google Cloud Translate
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter

# Utilities
requests==2.31.0 